refactor(render-session): share DaemonClientRenderSession delegate
Consolidates the ~150 LOC of pass-through `RenderSession` delegate
methods that previously lived in two copies (`SubprocessRenderSession`
in `:render-session-subprocess`, `EmbeddedDesktopRenderSession` in
`:render-session-embedded-desktop`) into a single public
`DaemonClientRenderSession` class living in `:render-session-subprocess`.

The class is transport-agnostic — it accepts a `DaemonClient`, a
`NotificationFanout`, and a `closeAction` lambda. Each backend's factory
(`SubprocessRenderSessions.open`, `EmbeddedDesktopRenderSessions.open`)
constructs one with its own close semantics:

- Subprocess: `closeAction = { spawn.shutdown(); fanout.clear() }`
- Embedded: `closeAction = { client.shutdownAndExit(); thread.join();
  pipes.close(); restoreSysprops() }`

`:render-session-embedded-desktop` gains an `implementation` dep on
`:render-session-subprocess` to reach the delegate. The third copy that
lives in `:mcp` (the supervisor's `SupervisedDaemon.session` view) stays
where it is — having `:mcp` depend on `:render-session-subprocess`
would create a cycle (subprocess already depends on `:mcp` for
`DaemonClient`). Lifting `DaemonClient` out of `:mcp` to break the
cycle is a substantial refactor across many files and out of scope.

Net: 2 copies → 1 for subprocess + embedded; `:mcp` still keeps its
own (slightly different — no-op close because supervisor owns
lifecycle). Future cleanup would lift `DaemonClient` into a neutral
module so all three backends can share.

## Test plan

- [x] `:render-session-subprocess:test` passes.
- [x] `:render-session-embedded-desktop:test` passes (including the
  end-to-end test that opens a real session against `:samples:cmp`).
- [x] `:mcp:test` passes.
- [x] `:cli:test` passes.
- [x] `ktfmtCheckAll` clean.
